Merge from Orgbook, Postgres branches

+++ agent/demo/alice-pg.py
@@ -16,10 +16,12 @@ run_mode = os.getenv('RUNMODE')
 
 internal_host = "127.0.0.1"
 external_host = "localhost"
+scripts_dir = "../scripts/"
 
 if run_mode == 'docker':
     internal_host = "host.docker.internal"
     external_host = "host.docker.internal"
+    scripts_dir = "scripts/"
 
 # some globals that are required by the hook code
 webhook_port = int(sys.argv[1])
@@ -45,6 +47,19 @@ class alice_webhooks(webhooks):
             print("#15 After receiving credential offer, send credential request")
             resp = requests.post(admin_url + '/credential_exchange/' + credential_exchange_id + '/send-request')
             assert resp.status_code == 200
+            return ""
+
+        elif state == 'stored':
+            print("Stored credential in wallet")
+            resp = requests.get(admin_url + '/credential/' + message['credential_id'])
+            assert resp.status_code == 200
+            print("Stored credential:")
+            print(resp.text)
+            print("credential_id", message['credential_id'])
+            print("credential_definition_id", message['credential_definition_id'])
+            print("schema_id", message['schema_id'])
+            print("credential_request_metadata", message['credential_request_metadata'])
+
             return ""
 
         return ""
@@ -103,7 +118,6 @@ def main():
     else:
         with open('local-genesis.txt', 'r') as genesis_file:
             genesis = genesis_file.read()
-    #print(genesis)
 
     # TODO seed from input parameter; optionally register the DID
     rand_name = str(random.randint(100000, 999999))
@@ -133,8 +147,9 @@ def main():
     wallet_key  = 'alice'+rand_name
     python_path = ".."
     webhook_url = "http://" + external_host + ':' + str(webhook_port) + "/webhooks"
-    (agent_proc, t1, t2) =  start_agent_subprocess(genesis, seed, endpoint_url, in_port_1, in_port_2, in_port_3, admin_port,
-                                            'indy', wallet_name, wallet_key, python_path, webhook_url)
+    (agent_proc, t1, t2) =  start_agent_subprocess('alice', genesis, seed, endpoint_url, in_port_1, in_port_2, in_port_3, admin_port,
+                                            'indy', wallet_name, wallet_key, python_path, webhook_url, 
+                                            scripts_dir, run_subprocess=True)
     time.sleep(3.0)
     print("Admin url is at:", admin_url)
     print("Endpoint url is at:", endpoint_url)
@@ -168,13 +183,14 @@ def main():
     except Exception as e:
         print(e)
     finally:
-        time.sleep(2.0)
-        agent_proc.terminate()
-        try:
-            agent_proc.wait(timeout=0.5)
-            print('== subprocess exited with rc =', agent_proc.returncode)
-        except subprocess.TimeoutExpired:
-            print('subprocess did not terminate in time')
+        if agent_proc:
+            time.sleep(2.0)
+            agent_proc.terminate()
+            try:
+                agent_proc.wait(timeout=0.5)
+                print('== subprocess exited with rc =', agent_proc.returncode)
+            except subprocess.TimeoutExpired:
+                print('subprocess did not terminate in time')
         sys.exit()
 
 if __name__ == "__main__":
